% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{calcular_votos_totales}
\alias{calcular_votos_totales}
\title{Title}
\usage{
calcular_votos_totales(bd, partido, eleccion, grupo = NULL)
}
\arguments{
\item{bd}{base de datos con resultados electorales}

\item{partido}{partidos o vector de partidos (ver base de datos de partidos)}

\item{eleccion}{elección o vector de elecciones de la cual se van a obtener los totales}

\item{grupo}{}
}
\value{
base de datos con cada una de las votaciones totales por partido en las elecciones solicitadas
}
\description{
Title
}
