% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{colorear_ganador_degradado}
\alias{colorear_ganador_degradado}
\title{Title}
\usage{
colorear_ganador_degradado(
  bd,
  eleccion,
  colores_nombrados,
  grupo,
  saturacion = 0.9
)
}
\arguments{
\item{saturacion}{}
}
\value{

}
\description{
Title
}
