% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_base_electoral.R
\name{detectar_partidos}
\alias{detectar_partidos}
\title{Title}
\usage{
detectar_partidos(bd)
}
\arguments{
\item{bd}{base de datos de cualquier elección en crudo (sin ningún procesamiento)}
}
\value{

}
\description{
Title
}
