% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_base_electoral.R
\name{repartir_coalicion}
\alias{repartir_coalicion}
\title{Title}
\usage{
repartir_coalicion(bd, identificadores, partido)
}
\arguments{
\item{bd}{base de datos de cualquier elección en crudo (sin ningún procesamiento)}

\item{identificadores}{nivel de observación para el que se aplica el análisis}

\item{partido}{los partidos identificados para cada elección. Calculados con la función detectar partidos}
}
\value{

}
\description{
Title
}
